+++ validity-prover/src/app/mod.rs
@@ -1,5 +1,6 @@
 pub mod check_point_store;
 pub mod error;
 pub mod observer;
-pub mod observer_sync;
+pub mod observer_view;
 pub mod validity_prover;
+pub mod validity_prover_view;
